Add a sparse version of the model, written in Cython
This version reads only the non-zero pixels of the data and calculates
predictions on those. Beucase the zero values are important for
caulculting Z (the partition function), we randomly sample zeros from
the matrix and use them too in the calculation.

The implementation also includes a pure python version for testing and
unit tests, comparing both the likelihood and gradients to the dense
version of the model.

Another feature that was added is checkpointing - each iteration can
save its result to a file and can resume from that point on the next
restart.

NOTE: The sparse model only works with mcool files with no duplicate
entries (i.e. more than one entry per pixel). We have code to remove
these duplicate entries, but it's slow and requires a lot of memory.