Add parallelisation to loglikelihood
Using Cython's prange (which needs OpenMP), we can split the calculation
to several threads and work on them in parallel. Use the 'nthreads'
kwarg for fit_sprase

For this I had to remove StreamingLogsumexp and replace it with a
struct+functions implementation. This allows us to malloc() it on demand
without using the GIL

Use all CPUs if nthreads is negative-like

Can simply done by passing 0 from the cli