Fix bug with override trans weights value
If trans shape is 'none' we would throw an error if the init_values
included a trans_weights element. The problem is that all our outputs
have this value, even if shape is 'none', since we copy it from the
cis_weights value. To work around this, we simply ignore the
trans_weights init_value